+++ processes/__init__.py
@@ -1,3 +1,4 @@
+from .covers import CoverProcess
 from .doab import DOABProcess
 from .gutenberg import GutenbergProcess
 from .hathiTrust import HathiTrustProcess